Comments and stdint

+++ mix.h
@@ -1,12 +1,19 @@
+/*
+ * Functions that convert the raw aileron, elevator, rudder, and brake
+ * inputs and convert them to outputs to the split elevons 
+ */
+
 #ifndef MIX_H
 #define MIX_H
 
+#include <Arduino.h>
+
 namespace Mix 
 {
-    int right_top(int arl, int ele, int rud, int brk);
-    int right_bottom(int arl, int ele, int rud, int brk);
-    int left_top(int arl, int ele, int rud, int brk);
-    int left_bottom(int arl, int ele, int rud, int brk);
+    uint16_t right_top(uint16_t arl, uint16_t ele, uint16_t rud, uint16_t brk);
+    uint16_t right_bottom(uint16_t arl, uint16_t ele, uint16_t rud, uint16_t brk);
+    uint16_t left_top(uint16_t arl, uint16_t ele, uint16_t rud, uint16_t brk);
+    uint16_t left_bottom(uint16_t arl, uint16_t ele, uint16_t rud, uint16_t brk);
 }
 
 #endif
